Move a bunch of code around
