+++ src/interpret.jl
@@ -522,53 +522,6 @@ function handle_err(stack, frame, pc, err)
 end
 
 """
-    pc = finish!(stack, frame, pc=frame.pc[])
-
-Run `frame` until execution terminates. `pc` is the program counter for the final statement.
-`stack` controls call evaluation; `stack = Compiled()` evaluates :call expressions
-by normal dispatch, whereas a vector of `JuliaStackFrame`s will use recursive interpretation.
-
-If execution hits a breakpoint, then `pc` is a reference to the breakpoint. `stack[end]`,
-if not running in `Compiled()` mode, will contain the frame in which the breakpoint was hit.
-"""
-function finish!(stack, frame, pc::JuliaProgramCounter=frame.pc[], istoplevel::Bool=false)
-    local new_pc
-    while true
-        new_pc = _step_expr!(stack, frame, pc, istoplevel)
-        (new_pc == nothing || isa(new_pc, BreakpointRef)) && break
-        pc = new_pc
-        if shouldbreak(frame, pc)
-            push!(stack, frame)
-            new_pc = BreakpointRef(frame.code, pc)
-            break
-        end
-    end
-    frame.pc[] = pc
-    return isa(new_pc, BreakpointRef) ? new_pc : pc
-end
-finish!(stack, frame, istoplevel::Bool) = finish!(stack, frame, frame.pc[], istoplevel)
-
-"""
-    ret = finish_and_return!(stack, frame, istoplevel::Bool=false)
-    ret = finish_and_return!(stack, frame, pc, istoplevel::Bool)
-
-Run `frame` until execution terminates, and pass back the computed return value.
-`stack` controls call evaluation; `stack = Compiled()` evaluates :call expressions
-by normal dispatch, whereas a vector of `JuliaStackFrame`s will use recursive interpretation.
-
-If execution hits a breakpoint, then `ret` is a reference to the breakpoint. `stack[end]`,
-if not running in `Compiled()` mode, will contain the frame in which the breakpoint was hit.
-
-Optionally supply the starting `pc`, if you don't want to start at the current location in `frame`.
-"""
-function finish_and_return!(stack, frame, pc::JuliaProgramCounter=frame.pc[], istoplevel::Bool=false)
-    pc = finish!(stack, frame, pc, istoplevel)
-    isa(pc, BreakpointRef) && return pc
-    return get_return(frame, pc)
-end
-finish_and_return!(stack, frame, istoplevel::Bool) = finish_and_return!(stack, frame, frame.pc[], istoplevel)
-
-"""
     ret = get_return(frame, pc=frame.pc[])
 
 Get the return value of `frame`. Throws an error if `pc` does not point to a `return` expression.
@@ -581,222 +534,3 @@ function get_return(frame, pc = frame.pc[])
     return @lookup(frame, (node::Expr).args[1])
 end
 get_return(frame::Tuple{Module,Expr,JuliaStackFrame}) = get_return(frame[end])
-
-"""
-    ret = finish_stack!(stack)
-
-Completely unwind `stack`, finishing it frame-by-frame. If execution hits a breakpoint,
-`ret` will be a reference to the breakpoint.
-"""
-function finish_stack!(stack)
-    while !isempty(stack)
-        frame = pop!(stack)
-        ret = finish_and_return!(stack, frame)
-        isa(ret, BreakpointRef) && return ret
-        isempty(stack) && return ret
-        parentframe = stack[end]
-        pc = parentframe.pc[]
-        if isassign(parentframe, pc)
-            lhs = getlhs(pc)
-            do_assignment!(parentframe, lhs, ret)
-        end
-        pc += 1
-        parentframe.pc[] = pc
-        shouldbreak(parentframe) && return BreakpointRef(parentframe.code, pc)
-    end
-    error("stack is empty")
-end
-
-function is_call(node)
-    isexpr(node, :call) ||
-    (isexpr(node, :(=)) && (isexpr(node.args[2], :call)))
-end
-
-"""
-    next_until!(predicate, stack, frame, pc=frame.pc[])
-
-Step through statements of `frame` until the next statement satifies `predicate(stmt)`.
-"""
-function next_until!(f, stack, frame, pc::JuliaProgramCounter=frame.pc[], istoplevel::Bool=false)
-    while (newpc = _step_expr!(stack, frame, pc, istoplevel)) != nothing
-        pc = newpc
-        if f(pc_expr(frame, pc)) || shouldbreak(frame, pc)
-            frame.pc[] = pc
-            return pc
-        end
-    end
-    frame.pc[] = pc
-    return nothing
-end
-next_until!(f, stack, frame, istoplevel::Bool) = next_until!(f, stack, frame, frame.pc[], istoplevel)
-next_call!(stack, frame, pc=frame.pc[]) = next_until!(node->is_call(node)||isexpr(node,:return), stack, frame, pc)
-
-"""
-    through_methoddef_or_done!(stack, frame)
-
-Runs `frame` at top level until it either finishes (e.g., hits a `return` statement)
-or defines a new method.
-"""
-function through_methoddef_or_done!(stack, frame::JuliaStackFrame)
-    predicate(stmt) = isexpr(stmt, :method, 3) || isexpr(stmt, :thunk)
-    pc = next_until!(predicate, stack, frame, true)
-    pc === nothing && return nothing
-    stmt = pc_expr(frame, pc)
-    return isexpr(stmt, :method, 3) ? #= normal exit =# _step_expr!(stack, frame, pc, true) :
-                                      #= breakpoint exit =# pc
-end
-through_methoddef_or_done!(stack, frame::Tuple{Module,Expr,JuliaStackFrame}) =
-    through_methoddef_or_done!(stack, frame[end])
-through_methoddef_or_done!(stack, modex::Tuple{Module,Expr,Expr}) = Core.eval(modex[1], modex[3])
-
-function changed_line!(expr, line, fls)
-    if length(fls) == 1 && isa(expr, LineNumberNode)
-        return expr.line != line
-    elseif length(fls) == 1 && isa(expr, Expr) && isexpr(expr, :line)
-        return expr.args[1] != line
-    else
-        if is_loc_meta(expr, :pop_loc)
-            pop!(fls)
-        elseif is_loc_meta(expr, :push_loc)
-            push!(fls,(expr.args[2],0))
-        end
-        return false
-    end
-end
-
-"""
-Determine whether we are calling a function for which the current function
-is a wrapper (either because of optional arguments or becaue of keyword arguments).
-"""
-function iswrappercall(expr)
-    isexpr(expr, :(=)) && (expr = expr.args[2])
-    isexpr(expr, :call) && any(x->x==SlotNumber(1), expr.args)
-end
-
-pc_expr(frame, pc) = frame.code.code.code[pc.next_stmt]
-pc_expr(frame) = pc_expr(frame, frame.pc[])
-
-function find_used(code::CodeInfo)
-    used = BitSet()
-    stmts = code.code
-    for stmt in stmts
-        Core.Compiler.scan_ssa_use!(push!, used, stmt)
-        if isexpr(stmt, :struct_type)  # this one is missed
-            for a in stmt.args
-                Core.Compiler.scan_ssa_use!(push!, used, a)
-            end
-        end
-    end
-    return used
-end
-
-isgotonode(node) = isa(node, GotoNode) || isexpr(node, :gotoifnot)
-
-"""
-    loc = whereis(frame, pc=frame.pc[])
-
-Return the file and line number for `frame` at `pc`.  If this cannot be
-determined, `loc == nothing`. Otherwise `loc == (filepath, line)`.
-
-When `frame` represents top-level code,
-"""
-function CodeTracking.whereis(framecode::JuliaFrameCode, pc)
-    codeloc = codelocation(framecode.code, convert(Int, pc))
-    codeloc == 0 && return nothing
-    lineinfo = framecode.code.linetable[codeloc]
-    return framecode.scope isa Method ?
-        whereis(lineinfo, framecode.scope) : string(lineinfo.file), lineinfo.line
-end
-CodeTracking.whereis(frame::JuliaStackFrame, pc=frame.pc[]) = whereis(frame.code, pc)
-
-# Note: linenumber is now an internal method for use by `next_line!`
-# If you want to know the actual line number in a file, call `whereis`.
-function linenumber(framecode::JuliaFrameCode, pc)
-    codeloc = codelocation(framecode.code, convert(Int, pc))
-    codeloc == 0 && return nothing
-    return framecode.code.linetable[codeloc].line
-end
-linenumber(frame::JuliaStackFrame, pc=frame.pc[]) = linenumber(frame.code, pc)
-
-function codelocation(code::CodeInfo, idx)
-    codeloc = code.codelocs[idx]
-    while codeloc == 0 && code.code[idx] === nothing && idx < length(code.code)
-        idx += 1
-        codeloc = code.codelocs[idx]
-    end
-    return codeloc
-end
-
-"""
-    stmtidx = statementnumber(frame, line)
-
-Return the index of the first statement in `frame`'s `CodeInfo` that corresponds to `line`.
-"""
-function statementnumber(framecode::JuliaFrameCode, line)
-    lineidx = searchsortedfirst(framecode.code.linetable, line; by=lin->isa(lin,Integer) ? lin : lin.line)
-    1 <= lineidx <= length(framecode.code.linetable) || throw(ArgumentError("line $line not found in $(framecode.scope)"))
-    return searchsortedfirst(framecode.code.codelocs, lineidx)
-end
-statementnumber(frame::JuliaStackFrame, line) = statementnumber(frame.code, line)
-
-"""
-    framecode, stmtidx = statementnumber(method, line)
-
-Return the index of the first statement in `framecode` that corresponds to the given `line` in `method`.
-"""
-function statementnumber(method::Method, line; line1=whereis(method)[2])
-    linec = line - line1 + method.line  # line number at time of compilation
-    framecode = get_framecode(method)
-    return framecode, statementnumber(framecode, linec)
-end
-
-function next_line!(stack, frame, dbstack = nothing)
-    initial = linenumber(frame)
-    first = true
-    pc = frame.pc[]
-    while linenumber(frame, pc) == initial
-        # If this is a return node, interrupt execution. This is the same
-        # special case as in `s`.
-        expr = pc_expr(frame, pc)
-        (!first && isexpr(expr, :return)) && return pc
-        first = false
-        # If this is a goto node, step it and reevaluate
-        if isgotonode(expr)
-            pc = _step_expr!(stack, frame, pc)
-            (pc == nothing || isa(pc, BreakpointRef)) && return pc
-        elseif dbstack !== nothing && iswrappercall(expr)
-            # With splatting it can happen that we do something like ssa = tuple(#self#), _apply(ssa), which
-            # confuses the logic here, just step into the first call that's not a builtin
-            while true
-                dbstack[end] = JuliaStackFrame(JuliaFrameCode(frame.code; wrapper = true), frame, pc)
-                call_expr = pc_expr(frame, pc)
-                isexpr(call_expr, :(=)) && (call_expr = call_expr.args[2])
-                call_expr = Expr(:call, map(x->@lookup(frame, x), call_expr.args)...)
-                new_frame = enter_call_expr(call_expr)
-                if new_frame !== nothing
-                    push!(dbstack, new_frame)
-                    frame = new_frame
-                    pc = frame.pc[]
-                    break
-                else
-                    pc = _step_expr!(stack, frame, pc)
-                    (pc == nothing || isa(pc, BreakpointRef)) && return pc
-                end
-            end
-        else
-            pc = _step_expr!(stack, frame, pc)
-            (pc == nothing || isa(pc, BreakpointRef)) && return pc
-        end
-        frame.pc[] = pc
-        shouldbreak(frame, pc) && return BreakpointRef(frame.code, pc)
-    end
-    maybe_next_call!(stack, frame, pc)
-end
-
-function maybe_next_call!(stack, frame, pc)
-    call_or_return(node) = is_call(node) || isexpr(node, :return)
-    call_or_return(pc_expr(frame, pc)) ||
-        (pc = next_until!(call_or_return, stack, frame, pc, false))
-    pc
-end
-maybe_next_call!(stack, frame) = maybe_next_call!(stack, frame, frame.pc[])
